backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "faker>=37.8.0",
    "uvicorn>=0.37.0",
]
